Moved constants.py up a level
