+++ dbt_platform_helper/utils/validation.py
@@ -12,10 +12,10 @@ from schema import Schema
 from schema import SchemaError
 from yaml.parser import ParserError
 
+from dbt_platform_helper.constants import CODEBASE_PIPELINES_KEY
+from dbt_platform_helper.constants import ENVIRONMENTS_KEY
+from dbt_platform_helper.constants import PLATFORM_CONFIG_FILE
 from dbt_platform_helper.utils.aws import get_aws_session_or_abort
-from dbt_platform_helper.utils.constants import CODEBASE_PIPELINES_KEY
-from dbt_platform_helper.utils.constants import ENVIRONMENTS_KEY
-from dbt_platform_helper.utils.constants import PLATFORM_CONFIG_FILE
 from dbt_platform_helper.utils.files import apply_environment_defaults
 from dbt_platform_helper.utils.messages import abort_with_error
