+++ tests/platform_helper/utils/test_files.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 import pytest
 
-from dbt_platform_helper.utils.constants import PLATFORM_CONFIG_FILE
+from dbt_platform_helper.constants import PLATFORM_CONFIG_FILE
 from dbt_platform_helper.utils.files import apply_environment_defaults
 from dbt_platform_helper.utils.files import generate_override_files
 from dbt_platform_helper.utils.files import is_terraform_project
